Split JsonSerializerGenerator into its own file
